+++ sinwan-test-app/src/main.tsx
@@ -1,0 +1,5 @@
+import { mount } from "sinwan/renderer";
+import { App } from "./App";
+
+import.meta.hot.accept();
+mount(App, document.getElementById("app")!);